Extract TemplateSubstitute to its own util
